+++ RandUsers/RandUsers/Data/Responses/ResultsResponse.swift
@@ -1,0 +1,13 @@
+//
+//  ResultsResponse.swift
+//  RandUsers
+//
+//  Created by Guido Fabio on 7/3/25.
+//
+
+import Foundation
+
+struct Results: Codable {
+    let info: InfoResponse
+    let results: [UserResponse]?
+}
